+++ examples/with-chi/main.go
@@ -88,6 +88,10 @@ func main() {
 						ClientID:     "4398792-907871294886928395",
 						ClientSecret: "His4yXGEovVp5TZkZhEAt0ZXGh8uOVDm",
 					}),
+					thirdparty.GoogleWorkspaces(tpmodels.GoogleWorkspacesConfig{
+						ClientID:     "1060725074195-kmeum4crr01uirfl2op9kd5acmi9jutn.apps.googleusercontent.com",
+						ClientSecret: "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",
+					}),
 				},
 			}),
 			session.Init(nil),