+++ recipe/thirdparty/tpmodels/apiInterface.go
@@ -39,7 +39,7 @@ type SignInUpPOSTResponse struct {
 		AuthCodeResponse interface{}
 	}
 	NoEmailGivenByProviderError *struct{}
-	FieldError                  *struct{ Error string }
+	FieldError                  *struct{ ErrorMsg string }
 }
 
 type APIOptions struct {