+++ recipe/thirdpartyemailpassword/tpepmodels/recipeInterface.go
@@ -34,7 +34,7 @@ type SignInUpResponse struct {
 		CreatedNewUser bool
 		User           User
 	}
-	FieldError *struct{ Error string }
+	FieldError *struct{ ErrorMsg string }
 }
 
 type SignUpResponse struct {